+++ lib/src/data/drivers/local_auth_driver.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
 import 'package:local_auth/local_auth.dart' hide BiometricType;
 import 'package:local_auth_android/local_auth_android.dart' hide BiometricType;
 import 'package:local_auth_darwin/local_auth_darwin.dart' hide BiometricType;
@@ -36,7 +35,7 @@ class LocalAuthDriver extends ILocalAuthDriver {
         ],
       );
       return Right(response);
-    } on PlatformException catch (e) {
+    } catch (e) {
       return Left(Exception(e));
     }
   }
@@ -46,7 +45,7 @@ class LocalAuthDriver extends ILocalAuthDriver {
     try {
       final response = await _auth.getAvailableBiometrics();
       return response.map((e) => biometricTypeFromTyoe(e.name)).toList();
-    } on PlatformException catch (e) {
+    } catch (e) {
       debugPrint('$e');
       return [];
     }
@@ -56,7 +55,7 @@ class LocalAuthDriver extends ILocalAuthDriver {
   Future<Either<Exception, bool>> canCheckBiometrics() async {
     try {
       return Right(await _auth.canCheckBiometrics);
-    } on PlatformException catch (e) {
+    } catch (e) {
       return Left(Exception(e));
     }
   }
@@ -65,7 +64,7 @@ class LocalAuthDriver extends ILocalAuthDriver {
   Future<Either<Exception, bool>> isDeviceSupported() async {
     try {
       return Right(await _auth.isDeviceSupported());
-    } on PlatformException catch (e) {
+    } catch (e) {
       return Left(Exception(e));
     }
   }
@@ -75,7 +74,7 @@ class LocalAuthDriver extends ILocalAuthDriver {
     try {
       await _auth.stopAuthentication();
       return Right(unit);
-    } on PlatformException catch (e) {
+    } catch (e) {
       return Left(Exception(e));
     }
   }