+++ lib/src/infra/drivers/firebase/i_firebase_driver.dart
@@ -2,19 +2,6 @@ import '../../../domain/interfaces/either.dart';
 
 abstract class IFirebaseDriver {
   Future<Either<Exception, Unit>> init({
-    String apiKey = '',
-    String appId = '',
-    String messagingSenderId = '',
-    String projectId = '',
-    String? authDomain,
-    String? databaseURL,
-    String? storageBucket,
-    String? measurementId,
-    String? trackingId,
-    String? deepLinkURLScheme,
-    String? androidClientId,
-    String? iosClientId,
-    String? iosBundleId,
-    String? appGroupId,
+    Map<String, String?> config = const {},
   });
 }